Use Polyester's @batch to speedup nonlinear prediction.
On the full HAR example, a single pass to predict vertices took about
95 s before the change. We considered two strategies:

1. Parallelize `predict`: assumes number of samples is large.
The strategy computes `kernel(xi, x)` in batches and accumulated the
prediction via 5-arg multiplications. Balancing work is challenging in
this paradigm.

2. Parallelize `classify`: assumes number of predictions is large. A
pass to compute errors scales as `n_train * n_predict`, where `n_train`
counts the number of samples used to train the model and `n_testing` is
the number of samples we wish to predict on.

Strategy (2) seems to scale better and requires less manual tuning.
The change manages to cut down prediction in HAR from 192 s to <10 s.